Upper level drivers interface is defined, except lineread sorting.  Lower level (interaction with particular driver) still missing. dataformat.tli.txt file added for human readable tli.  magic digits corrected

+++ lineread/include/version_lr.h
@@ -24,8 +24,8 @@
 #define _VERSION_LR_H_
 
 
-int version = 4;
-int revision = 0;
+unsigned short version = 4;
+unsigned short revision = 0;
 int version_rc = -1;
 
 int TLIversion = 4;
